api-proxy/requirements.txt
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# Opcional: para logging, métricas, etc.
# loguru==0.7.2

# Google Cloud Services
google-cloud-texttospeech==4.15.0
google-auth==2.28.1
google-api-core==2.17.1

# FastAPI y componentes
fastapi==0.115.0
uvicorn==0.34.0
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic (versión v2 - más moderna y rápida)
pydantic==2.5.0
pydantic-core==2.14.0

# Flask (si todavía lo usas junto con FastAPI)
flask==3.0.0
flask-cors==4.0.0

# Utilidades y seguridad
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.8
pyjwt==2.8.0
bcrypt==4.1.2

# Procesamiento de documentos
python-docx==1.1.0
lxml==4.9.4

# Utilidades del sistema
click==8.1.7
httpx==0.25.2